# phase3_django_api/library_management_system/swagger/schema.yaml
openapi: 3.0.3
info:
  title: ''
  version: 0.0.0
paths:
...
  /api/v1/borrowings/:
    get:
      operationId: api_v1_borrowings_list
      parameters:
      - in: query
        name: book
        schema:
          type: integer
      - in: query
        name: borrow_date_after
        schema:
          type: string
          format: date
      - in: query
        name: borrow_date_before
        schema:
          type: string
          format: date
      - in: query
        name: due_date_after
        schema:
          type: string
          format: date
      - in: query
        name: due_date_before
        schema:
          type: string
          format: date
      - in: query
        name: late_fee__gte
        schema:
          type: number
      - in: query
        name: late_fee__lte
        schema:
          type: number
      - in: query
        name: member
        schema:
          type: integer
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - in: query
        name: return_date
        schema:
          type: string
          format: date
      - in: query
        name: return_date_isnull
        schema:
          type: boolean
      tags:
      - api
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedBorrowingList'
          description: ''
    post:
      operationId: api_v1_borrowings_create
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Borrowing'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Borrowing'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Borrowing'
        required: true
      security:
      - cookieAuth: []
      - basicAuth: []
      - {}
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Borrowing'
          description: ''
...
components:
  schemas:
...
    Borrowing:
      type: object
      properties:
        borrowing_id:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
        borrow_date:
          type: string
          format: date
        due_date:
          type: string
          format: date
        return_date:
          type: string
          format: date
          nullable: true
        late_fee:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        member:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
        book:
          type: integer
      required:
      - book
      - borrow_date
      - borrowing_id
      - created_at
      - due_date
      - late_fee
      - member
      - updated_at
...
    PaginatedBorrowingList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Borrowing'
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid